// functions/src/cloud_functions/bets/models/bets/paralympics_bet.ts
export interface ParalympicsBet {
  dateTime: Date;
  gameId: string;
  week: string;
  rivalName: string;
  playerName: string;
  league: string;
  betTeam: string;
  eventType: string;
  clientVersion: string;
  rivalCountry: string;
  playerCountry: string;
  betProfit: number;
  uid: string;
  betAmount: number;
  winner: string;
  gameStartDateTime: string;
  gameName: string;
  isClosed: boolean;
  dataProvider: string;
  id: string;
  event: string;
  username: string;
}